% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_survival.R
\name{autoplot.grouped_marginal_survival}
\alias{autoplot.grouped_marginal_survival}
\title{Plot of marginal survival curve by group}
\usage{
\method{autoplot}{grouped_marginal_survival}(object, xlab = "Time", strata_labels = c("Control", "Treated"), ...)
}
\arguments{
\item{object}{An object of class \code{grouped_marginal_survival}.}

\item{xlab}{Label for the x-axis.}

\item{strata_labels}{Labels for each level of the stratification variable.}

\item{...}{Currently unused.}
}
\value{
A \code{\link{ggplot2}} object.
}
\description{
Quickly plot a marginal survival using \code{\link{ggplot2}} with grouped data. Survival
curves for each group are stratified by a stratification variable and plots
are faceted by group.
}
