% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{autoplot.missing_counts}
\alias{autoplot.missing_counts}
\title{Plot missing observations}
\usage{
\method{autoplot}{missing_counts}(object, yvar = c("prop_missing", "n_missing"))
}
\arguments{
\item{object}{An object of class \code{\link[=count_missing]{missing_counts}}.}

\item{yvar}{Character vector of length 1 indicating whether the y-axis should
report the proportion of missing observations (\code{"prop_missing"}) or the
number of missing observations (\code{"n_missing"}).}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
Bar plot to summarize the number of missing observations for each variable
selected in \code{\link[=count_missing]{count_missing()}}. Plots are faceted by analysis number and
variables are on the x-axis in each facet.
}
