% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{autoplot.observed_vs_imputed}
\alias{autoplot.observed_vs_imputed}
\title{Observed vs. imputed plot}
\usage{
\method{autoplot}{observed_vs_imputed}(object)
}
\arguments{
\item{object}{An \code{\link{observed_vs_imputed}} object.}
}
\value{
A \link[ggplot2:ggplot]{ggplot2::ggplot} object.
}
\description{
A plot comparing observed and imputed data for each variable. There is currently
assumed to be a group column (\code{grp}) which identifies the separate analyses.
The plot uses \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} with groups as rows and variables
as columns. Categorical data is plotted with a bar plot and kernel density
estimates are used for the continuous variables.
}
