% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmark_hazard_ratios.R
\name{benchmark_hazard_ratios}
\alias{benchmark_hazard_ratios}
\alias{autoplot.benchmarked_hazard_ratios}
\alias{html_table.benchmarked_hazard_ratios}
\title{Benchmark hazard ratios}
\usage{
benchmark_hazard_ratios(
  object,
  benchmarks = NULL,
  benchmark_estimate = "estimate",
  benchmark_lower = "lower",
  benchmark_upper = "upper"
)

\method{autoplot}{benchmarked_hazard_ratios}(
  object,
  xlab = "Benchmark hazard ratio",
  ylab = "Estimated hazard ratio",
  scale_breaks = seq(0.4, 2, 0.2),
  ...
)

\method{html_table}{benchmarked_hazard_ratios}(x, ...)
}
\arguments{
\item{object}{A \code{\link{surv_ate}} object.}

\item{benchmarks}{A data frame containing benchmark hazard ratios, potentially
by group. It must}

\item{benchmark_lower}{The column identifier for the lower confidence limit
of the benchmark hazard ratio.}

\item{benchmark_upper}{The column identifier for the upper confidence limit
of the benchmark hazard ratio.}

\item{becnhmark_estimate}{The column identifier for the benchmark hazard ratio
estimate.}
}
\value{
A \code{hazard_ratio} object that inherits from \code{\link[dplyr:reexports]{dplyr::tibble}}.
}
\description{
Obtain estimated hazard ratios in a tidy data frame from a \code{surv_ate} object and
compare them to benchmark hazard ratios (e.g., from target randomized clinical
trials). Quickly plot comparisons of the estimated hazard ratios to the benchmarks
using \code{autoplot()}. Generate a pretty HTML table with 95 percent confidence
intervals using \code{html_table()}.
}
