% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{count_by}
\alias{count_by}
\title{Counts by group}
\usage{
count_by(data, vars, by, max_n = Inf)
}
\arguments{
\item{data}{A dataset returned by \code{\link[ecdata:pin_disease]{ecdata::pin_nsclc()}} and preprocessed
using \code{\link[=preprocess]{preprocess()}}.}

\item{vars}{A character vector indicating the variables to count unique
values for.}

\item{by}{Columns to group by.}

\item{max_n}{The maximum number of values in a group in the returned data.}
}
\value{
A \code{\link[dplyr:reexports]{dplyr::tibble}}.
}
\description{
Count the number of unique values of one ore more variables by groups.
}
