% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{count_missing}
\alias{count_missing}
\title{Count missing observations}
\usage{
count_missing(data, vars)
}
\arguments{
\item{data}{A dataset returned by \code{\link[ecdata:pin_disease]{ecdata::pin_nsclc()}} and preprocessed
using \code{\link[=preprocess]{preprocess()}}.}

\item{vars}{A character vector of the names of variables to count the
number of missing observations for. If the vector is named, then the names
are used as labels for the variables.}
}
\value{
An object of class \code{missing_counts} inheriting from \code{\link[dplyr:reexports]{dplyr::tibble}}
with the following columns:
\describe{
\item{label}{The label for each variable in \code{vars}.}
\item{source_type}{Indicates whether the data is from the external control or the trial.}
\item{analysis_num}{An integer indicating the analyis number.}
\item{n_missing}{The number of missing observations.}
\item{n}{The number of total observations.}
\item{prop_missing}{The proportion of observations that are missing.}
}
}
\description{
Count missing observations for variables contained in a dataset.
}
