% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmeta.nsclc.R
\docType{package}
\name{ecmeta.nsclc-package}
\alias{ecmeta.nsclc}
\alias{ecmeta.nsclc-package}
\title{ecmeta.nsclc: Application of a Meta-Analytic Framework for Decision Making and Error Control to External Control Analyses in NSCLC}
\description{
A proof of concept in non-small cell lung cancer (NSCLC) of a
  meta-analytic framework to adjust external control studies for additional bias
  and variability due to the non-randomized design. A total of 14 pairwise comparisons
  are reanalyzed using propensity score methods as the basis for application of the
  framework.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.roche.com/incertid/ecmeta-manuscript/tree/master/ecmeta.nsclc}
}

}
\author{
\strong{Maintainer}: Devin Incerti \email{incerti.devin@gene.com}

Authors:
\itemize{
  \item Michael Bretscher \email{michael.bretscher@roche.com}
  \item Ray Lin \email{lins24@gene.com}
  \item Chris Harbron \email{chris.harbron@roche.com}
}

}
