% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecmeta_lpo_cv.R
\name{ecmeta_lpo_cv}
\alias{ecmeta_lpo_cv}
\title{Meta-analytic leave-p-out cross validation}
\usage{
ecmeta_lpo_cv(
  data,
  p = 1,
  method = "ml",
  ecmeta_args = NULL,
  predict_args = NULL
)
}
\arguments{
\item{data}{A data frame with columns \code{param}, \code{analysis_num},
\code{estimate}, and \code{se}. \code{param} should have three values representing the log
hazard ratio of a comparison between (i) the treatment and the external
control (\code{trt_ec}), (ii) the internal control and the external control
(\code{ic_ec}), and (iii) the treatment and the internal control (\code{trt_ic}).
\code{analysis_num} is the analysis (i.e., study) number, \code{estimate} is the point
estimate, and \code{se} is the standard error.}

\item{p}{Number of data points to use for testing. Default is 1 which is
leave-one-out cross validation.}

\item{method}{The method/software used to estimate the meta-analytic model
as in \code{ecmeta}.}

\item{ecmeta_args}{A named list containing additional arguments to pass
to \code{\link[ecmeta:ecmeta]{ecmeta::ecmeta()}}.}

\item{predict_args}{A named list containing additional arguments to pass to
\code{\link[ecmeta:predict.ecmeta]{ecmeta::predict.ecmeta()}}.}
}
\value{
A data frame where there are \code{np} rows (\code{n} = each time the model
is trained and tested; \code{p} is the number of testing observations for each
of the \code{n} iterations). Each row contains performance metrics.
}
\description{
Use leave-p-out cross-validation (LpO CV) to compare an external control
analysis approach that uses the \code{ecmeta} package to adjust for bias to a
standard approach that makes no adjustments. LpO CV uses p data points in the
data for testing and all other data points for training.
}
\details{
LpO CV requires training and validating
the dataset \eqn{C_p^n} where \eqn{n} is the number of observations in the original dataset.
For each of the \eqn{C_p^n} iterations, a meta-analytic model is fit to the training
data using \code{ecmeta::ecmeta()} and a prediction on the training set is made with
\code{ecmeta::predict.ecmeta()}. Predictions that are adjusted with the meta-analytic method
are compared to those without using any adjustment.
}
