% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ps.R
\name{fit_ps}
\alias{fit_ps}
\alias{fit_ps_mi}
\title{Fit propensity score model}
\usage{
fit_ps(formula, data, engine = "glm", ...)

fit_ps_mi(formula, data, engine = "glm", ...)
}
\arguments{
\item{formula}{An object of class \code{\link{formula}} with the treatment variable to the
left of the \code{~} operator and the model terms on the right.}

\item{data}{A data frame containing the variables in the model.
For \code{fit_ps_mi()}, there must be an additional column \code{imp} denoting the
imputation number.}

\item{engine}{The "engine" used to fit the propensity score model. Currently
limited to logistic regression fit with \code{\link[stats:glm]{stats::glm()}}.}

\item{...}{Additional arguments to pass to \code{\link[stats:glm]{stats::glm()}}.}
}
\value{
\code{fit_ps()} returns an object of class \code{fit_ps} with an element \code{fit} containing
the underlying fitted model. \code{fit_ps_mi()} returns an object of class\code{fit_ps_mi},
which is a list of \code{fit_ps} objects.
}
\description{
Fit a propensity score model using a specified engine. \code{fit_ps()} fits
a model on a single dataset and \code{fit_ps_mi()} fits a model on each of
\code{m} multiply imputed datasets.
}
