% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hazard_ratio.R
\name{hazard_ratio}
\alias{hazard_ratio}
\alias{autoplot.hazard_ratio}
\title{Hazard ratios}
\usage{
hazard_ratio(object, exponentiate = TRUE)

\method{autoplot}{hazard_ratio}(object)
}
\arguments{
\item{object}{A \code{\link{surv_ate}} object.}

\item{exponentiate}{Logical indicating whether or not to exponentiate
the coefficient estimates. Default is \code{TRUE} meaning that hazard ratios
are returned; if \code{FALSE}, then log hazard ratios are returned.}
}
\value{
A \code{hazard_ratio} object that inherits from \code{\link[dplyr:reexports]{dplyr::tibble}}.
}
\description{
Obtain estimated hazard ratios in a tidy data frame from a \code{surv_ate} object and
quickly plot them using \code{autoplot()}.
}
