% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{html_table.analysis}
\alias{html_table.analysis}
\title{HTML analysis table}
\usage{
\method{html_table}{analysis}(x)
}
\description{
Create an HTML table using using \code{\link[knitr:kable]{knitr::kable()}} (and styled with \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}})
that summarizes an \code{\link[=new_analysis]{analysis}} object. For each analysis, it indicates
the external control data source, the clinicaltrials.gov identifier for the clinical trial,
the experimental and comparator arm, and the sample size for both the external control and
the experimental arm of the clinical trial.
}
