% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{html_table}
\alias{html_table}
\alias{html_table.default}
\title{HTML table}
\usage{
html_table(x, ...)

\method{html_table}{default}(x)
}
\description{
Create an HTML table from an \code{R} object using \code{\link[knitr:kable]{knitr::kable()}} and style it
using \code{\link[kableExtra:kable_styling]{kableExtra::kable_styling()}}.
}
\seealso{
\code{\link[=html_table.analysis]{html_table.analysis()}}
}
