% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{impute_formula}
\alias{impute_formula}
\alias{add_impute_formula}
\title{Formulas for multiple imputation}
\usage{
impute_formula(formula, interaction = TRUE)

add_impute_formula(x, formula, interaction = TRUE)
}
\arguments{
\item{formula}{A \code{\link{formula}} object in \code{impute_formula} and a column
identifier for the formula objects in \code{add_impute_formula()}.}

\item{interaction}{Logical. If \code{TRUE}, then treatment status is interacted with
each covariate, which results in separate imputations models for the trial
and external control.}

\item{x}{An \code{\link[=new_analysis]{analysis}} object.}
}
\value{
\code{impute_formula()} returns a single \code{\link{formula}} object.
\code{add_impute_formuls} returns an \code{\link[=new_analysis]{analysis}} object with a
new list column named \code{impute_formula}, where each element is the formula used
for imputation of covariates required for a particular pairwise analysis.
}
\description{
Create formulas for multiple imputation. \code{impute_formula()} creates the
imputation formulas as functions of an existing formula object (e.g,
a formula used for a propensity score model) and \code{add_impute_formula()} adds
them to an \code{\link[=new_analysis]{analysis}}
object.
}
\seealso{
\code{\link{psmodel_specification}}
}
