% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_survival.R
\name{interpolate_surv}
\alias{interpolate_surv}
\title{Interpolate survival curves}
\usage{
interpolate_surv(x, times, extend = FALSE)
}
\arguments{
\item{x}{Survival curves. Assumed that there are columns for \code{strata},
\code{time}, and \code{surv}.}

\item{times}{Times at which to subset survival curves to.}

\item{extend}{Whether to extend survival predictions beyond maximum followup
time.}
}
\description{
Interpolate survival curves to obtain survival probabilities at specific times,
which may differ from the times at which the original curves were evaluated.
The step-function nature of the
predictions are preserved so that if \code{times} lies within \verb{[t1, t2)}, the
survival curves evaluated at \code{t1} will be used. An efficient approach using
\code{data.table} is employed.
}
