% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{label_ps_method}
\alias{label_ps_method}
\title{Labels for propensity score methods}
\usage{
label_ps_method(x)
}
\arguments{
\item{x}{A variable containing the propensity methods.}
}
\value{
A \code{\link{factor}} variable.
}
\description{
Convert a variable containing propensity score methods (as in the argument
\code{methods} from \code{\link[=psweight]{psweight()}}) to a factor variable with pretty labels for
the propensity score methods.
}
\examples{
label_ps_method(c("iptw_att", "iptw_att_trim", 
                  "match_nearest", "match_nearest_caliper",
                  "match_genetic", "match_genetic_caliper", 
                  "ps_stratification", "unadjusted"))
}
