% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_xdata.R
\name{make_xdata}
\alias{make_xdata}
\title{Model data}
\usage{
make_xdata(
  formula,
  data,
  imputation = c("multi", "none"),
  missing_dummy = FALSE,
  id = "patient_id",
  n_impute = 5
)
}
\arguments{
\item{formula}{A formula for the propensity score with the treatment variable
to the left of the \code{~} operator and the model terms on the right.}

\item{data}{The dataset containing the variables specified in \code{formula}.}

\item{imputation}{Imputation strategy to use. Use \code{"multi"} to multiply impute
data with \code{\link[=multi_impute]{multi_impute()}}. No imputation is performed when the \code{"none"} is
chosen.}

\item{missing_dummy}{Whether to create dummy variables for variables
when observations are missing.}

\item{id}{The variable uniquely identifying each patient.}

\item{n_impute}{Number of imnputations to perform. Only used when
\code{imputation = "multi"}.}
}
\value{
An object of class \code{model_data} inheriting from \code{\link[dplyr:reexports]{dplyr::tibble}}.
It contains all variables used in \code{formula} and, if multiple imputation
is used, an additional column named \code{imp} indicating the imputation
number.
}
\description{
Create data containing the treatment variable and all other covariates. This
data is required to build the design (i.e., \code{x}) matrix in either a propensity
score model or a covariate adjusted outcomes model. Only variables required
by the model are kept.
}
