% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_survival.R
\name{marginal_survival}
\alias{marginal_survival}
\alias{marginal_survival.psweight}
\alias{marginal_survival.psweight_mi}
\alias{marginal_survival.survfit}
\title{Marginal survival curves}
\usage{
marginal_survival(object, ...)

\method{marginal_survival}{psweight}(object, formula, times = NULL, ...)

\method{marginal_survival}{psweight_mi}(object, formula, times = NULL, ...)

\method{marginal_survival}{survfit}(object, times = NULL, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{...}{Currently unused.}

\item{formula}{A \code{\link{formula}} object as in \code{\link[survival:survfit]{survival::survfit}}. The right
hand side of the formula should have a single \code{strata} term.}

\item{times}{Unique times to compute survival probabilities at.}
}
\value{
An object of class \code{marginal_survival}, which is a
\code{\link[dplyr:reexports]{dplyr::tibble}} containing the following columns:

\describe{
\item{\{\code{strata}\}}{The name of the stratification variable. This will be the same
as the name of the variable supplied by the user.}
\item{\code{time}}{The time at which survival was evaluated.}
\item{\code{surv}}{The survival probability.}
}
}
\description{
This is a generic method for estimating marginal survival functions by levels
of a stratification variable (typically treatment vs. control).
}
