% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{multi_impute}
\alias{multi_impute}
\title{Multiple imputation}
\usage{
multi_impute(object, data, id = "patient_id", ...)
}
\arguments{
\item{object}{An \code{\link[Hmisc:aregImpute]{Hmisc::aregImpute}} object.}

\item{data}{The dataset that will be multiply imputed.}

\item{id}{Name of the variable in \code{data} uniquely identifying each row.
This is added as a column to the output.}
}
\value{
An "stacked" \code{\link[dplyr:reexports]{dplyr::tibble}} containing all of the multiply
imputed datasets row binded together. A column \code{imp} identifies the
imputation number.
}
\description{
Multiple imputation of covariates for the propensity score model using
\code{\link[Hmisc:transcan]{Hmisc::impute.transcan()}}. The model used for multiple imputation must
be fit with \code{\link[Hmisc:aregImpute]{Hmisc::aregImpute()}}.
}
