% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{new_analysis}
\alias{new_analysis}
\title{New analysis object}
\usage{
new_analysis(data)
}
\arguments{
\item{data}{A dataset returned by \code{\link[ecdata:pin_disease]{ecdata::pin_nsclc()}} and preprocessed
using \code{\link[=preprocess]{preprocess()}}.}
}
\value{
An object of class \code{analysis} that inherits from \code{\link[dplyr:reexports]{dplyr::tibble}}.Each
row is a unique analysis.
}
\description{
Create a new analysis object that summarizes all pairwise analyses used for
this study. Each pairwise analysis is uniquely identified by the clinical trial
and the experimental vs. comparator arm.
}
