% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute.R
\name{observed_vs_imputed}
\alias{observed_vs_imputed}
\title{Observed vs imputed data}
\usage{
observed_vs_imputed(
  object,
  data,
  type = c("categorical", "continuous"),
  labels = NULL
)
}
\arguments{
\item{object}{An \code{\link[Hmisc:aregImpute]{Hmisc::aregImpute}} object.}

\item{data}{A dataset that was multiply imputed using \code{\link[=multi_impute]{multi_impute()}}.}

\item{type}{Whether to examine continuous or categorical data. Both cannot
be assessed at the same time since they are different data types.}

\item{labels}{Optional labels for the variables contained in the columns of
\code{data}.}
}
\value{
An \code{observed_vs_imputed} object that inherits from \code{\link[dplyr:reexports]{dplyr::tibble}}
containing the columns:
\describe{
\item{value}{The value of a particular variable.}
\item{type}{Either \code{"Observed"} or the imputation number (e.g., \code{"Imputation 1"},
\code{"Imputation 2"}, ...).}
\item{var}{The name of the variable that was imputed.}
\item{label}{A pretty label for the variable that was imputed.}
}
}
\description{
Compare the observed and imputed values of either all of the continuous
or all of the categorical variables used in an imputation.
}
