% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{plot_bar}
\alias{plot_bar}
\title{Bar plot for categorical variable}
\usage{
plot_bar(data, var)
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
Draw bar plots for a categorical variable faceted by the
analysis number.
}
