% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{plot_density.data.frame}
\alias{plot_density.data.frame}
\title{Density plot for continuous variable}
\usage{
\method{plot_density}{data.frame}(data, var, xlim = NULL)
}
\arguments{
\item{data}{A dataset returned by \code{\link[ecdata:pin_disease]{ecdata::pin_nsclc()}} and preprocessed
using \code{\link[=preprocess]{preprocess()}}.}

\item{var}{A character vector of  length 1 indicating the variable to plot.}

\item{xlim}{Limits for the x-axis passed to \code{\link[ggplot2:coord_cartesian]{ggplot2::coord_cartesian()}}.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
Plot kernel density estimates of a continuous variable faceted by the
analysis number.
}
