% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psweight.R
\name{plot_density.grouped_psweight_mi}
\alias{plot_density.grouped_psweight_mi}
\title{Density plots to check balance}
\usage{
\method{plot_density}{grouped_psweight_mi}(object, method = "iptw_att", var, xlab = NULL)
}
\description{
Check the balance of continuous covariates with density plots.
}
