% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_density}
\alias{plot_density}
\title{Plot density}
\usage{
plot_density(object, ...)
}
\description{
Generic methods for plotting kernel density estimates.
}
\seealso{
\code{\link[=plot_density.data.frame]{plot_density.data.frame()}}, \code{\link[=plot_density.grouped_psweight_mi]{plot_density.grouped_psweight_mi()}}
}
