% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psweight.R
\name{plot_ps}
\alias{plot_ps}
\alias{plot_ps.grouped_psweight_mi}
\title{Plot propensity scores}
\usage{
plot_ps(object, ...)

\method{plot_ps}{grouped_psweight_mi}(object, method = "iptw_att", logit = FALSE)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{method}{A character vector of length one indicating the propensity score
method used to generate the weights.}

\item{logit}{Logical. If \code{TRUE}, then the logit of the propensity score is
plotted.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
Kernel density estimates of the propensity scores by treatment assignment in
order to check whether the estimated propensity scores are well balanced.
If \code{object} is of class \code{grouped_psweight}, then density plots are faceted by
the group number. If \code{object} is of class \code{grouped_psweight_mi}, then propensity
scores from the imputations are pooled.
}
