% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psweight.R
\name{plot_weights}
\alias{plot_weights}
\alias{plot_weights.grouped_psweight_mi}
\title{Plot propensity score weights}
\usage{
plot_weights(object, ...)

\method{plot_weights}{grouped_psweight_mi}(object, method = "iptw_att")
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{method}{A character vector of length one indicating the propensity score
method used to generate the weights.}
}
\value{
A \code{\link[ggplot2:ggplot]{ggplot2::ggplot}} object.
}
\description{
A histogram displaying propensity score weights. If \code{object} is of class
\code{grouped_psweight}, then histograms are faceted by the group number. If
\code{object} is of class \code{grouped_psweight_mi}, then weights from the imputations
are pooled.
}
