% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_ps.R
\name{predict_ps}
\alias{predict_ps}
\alias{predict_ps.fit_ps}
\alias{predict_ps.fit_ps_mi}
\title{Predict propensity scores}
\usage{
predict_ps(object, ...)

\method{predict_ps}{fit_ps}(object, id = "patient_id", ...)

\method{predict_ps}{fit_ps_mi}(object, id = "patient_id", ...)
}
\arguments{
\item{object}{A fitted propensity score model of the appropriate class.}

\item{...}{Currently unused.}

\item{id}{Name of the variable from the data frame used to fit \code{object} that
uniquely identifies each row. This is added as a column to the output.}
}
\value{
An object of class \code{ps} or \code{ps_mi}. A \code{ps_mi} object is a list
of \code{ps} objects (one for each imputation) and a \code{ps} object is a list with
elements:
\describe{
\item{ps}{The propensity score.}
\item{lp}{The linear predictor (i.e., the logit of the propensity score).}
\item{data}{The data used to fit the propensity score model.}
\item{formula}{The model formula used to fit the propensity score model.}
\item{treat}{The name of the treatment variable.}
\item{x_vars}{The names of all covariates (prior to any transformations or
dummy variable coding) used in the propensity score model.}
\item{id}{The }
}
}
\description{
Predict propensity scores from a fitted statistical model. Fitted value from
the model fit are used.
}
