% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-prep.R
\name{preprocess}
\alias{preprocess}
\title{Preprocess data}
\usage{
preprocess(data, combine_levels = TRUE, drop_nos = TRUE)
}
\arguments{
\item{data}{The unprocessed \code{\link[ecdata:nsclc]{ecdata::nsclc}} dataset.}

\item{combine_levels}{If \code{TRUE}, then levels of factor variables with
low counts (typically less than 10) are combined into a single category.
This current only applies to the race variable.}

\item{drop_nos}{In some instance, patients with a not otherwise specified
(NOS) histology are part of the external control but not part of the trial.
If \code{TRUE}, then these patients are removed from the external control cohort;
if \code{FALSE}, then they are not.}
}
\value{
A \code{\link[dplyr:reexports]{dplyr::tibble}}.
}
\description{
Preprocess the external control datasets for propensity score modeling. This
involves combining (i.e., grouping) categorical variables, deriving continuous
variables (i.e., time since initial diagnosis), and identifying each of the
distinct analyses for the study (defined by a unique clinical trial and
pairwise comparison between an experimental and comparator arm).
}
