% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_formula.R
\name{ps_formula}
\alias{ps_formula}
\alias{get_ps_formula}
\alias{add_ps_formula}
\title{Propensity score formulas}
\usage{
get_ps_formula(
  splines = TRUE,
  interaction = TRUE,
  treat = "treat",
  keep_names = FALSE
)

add_ps_formula(x, splines = TRUE, interaction = TRUE)
}
\arguments{
\item{splines}{Whether to use natural cubic splines (\code{\link[splines:ns]{splines::ns()}}) for
continuous variables.}

\item{interaction}{Whether to include an interaction between cancer stage
and days since diagnosis.}

\item{treat}{Name of the treatment variable. Default is \code{"treat"}.}

\item{keep_names}{If \code{TRUE}, then \code{analysis_id} is used to name the list.
Otherwise an unnamed list is returned.}

\item{x}{An \code{\link[=new_analysis]{analysis}} object.}
}
\value{
\code{get_ps_formula()} returns a list of formulas. \code{add_ps_formula()}
returns an \code{\link[=new_analysis]{analysis}} object
with a new list column named \code{ps_formula} where each element is the formula
used for a particular pairwise analysis.
}
\description{
Formulas for modeling the propensity score. \code{get_ps_formula()} obtains
the formulas and \code{add_ps_formula()} adds them to an existing
\code{\link[=new_analysis]{analysis}} object.
}
\seealso{
\code{\link{psmodel_specification}}
}
