% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ps_surv.R
\name{ps_surv}
\alias{ps_surv}
\alias{map_ps_surv}
\title{Propensity score survival analysis}
\usage{
ps_surv(xdata, formula = NULL, ydata, response, id = "patient_id", ...)

map_ps_surv(
  xdata_list,
  formula = NULL,
  ydata_list,
  response,
  id = "patient_id",
  grp_id,
  integer_grp_id = FALSE,
  ...
)
}
\arguments{
\item{xdata, xdata_list}{A dataset containing the variables for propensity
score modeling. This can either be an object produced by \code{make_xdata()}
or a standard data frame. If a standard data frame, then \code{formuls} cannot be
\code{NULL} and \code{make_xdata()} is run using \code{formula}.}

\item{formula}{The propensity score formula as described in \code{\link[=fit_ps]{fit_ps()}}. Only
required if \code{data} was not produced by \code{make_xdata()}.}

\item{ydata, ydata_list}{A data frame or list of data frames containing the
response variables as described in \code{\link[=surv_ate]{surv_ate()}}.}

\item{id}{The name of the column in \code{xdata} and \code{ydata} identifying each patient.}

\item{...}{Additional arguments to pass to \code{make_xdata()}.}

\item{A}{string containing the response for a survival model as described in
\code{\link[=surv_ate]{surv_ate()}}.}
}
\value{
\code{ps_surv()} returns an object of class \code{ps_surv} and \code{map_ps_surv()}
returnes an object of class \code{grouped_ps_surv}. Both are lists containing
the elements:
\describe{
\item{psfit}{The propensity score model fit using \code{\link[=fit_ps]{fit_ps()}} or \code{\link[=fit_ps_mi]{fit_ps_mi()}}.}
\item{psweight}{The propensity score weights returned by \code{\link[=psweight]{psweight()}}.}
\item{ate}{The average treatment effects returned by \code{\link[=surv_ate]{surv_ate()}}.}
\item{call}{The function call returned by \code{\link[=match.call]{match.call()}}.}
}
}
\description{
Run a complete pipeline required to estimate average treatment effects (ATEs) with
survival data using propensity score methods. The pipeline
(1) fits the propensity score model with \code{\link[=fit_ps]{fit_ps()}} or \code{\link[=fit_ps_mi]{fit_ps_mi()}},
(2) predict the propensity score with \code{\link[=predict_ps]{predict_ps()}}, (3) weights
patients using the propensity scores with \code{\link[=psweight]{psweight()}}, and estimates
ATEs with \code{\link[=surv_ate]{surv_ate()}}. \code{ps_surv()} implements the pipeline for a single group
while \code{map_ps_surv()} iterates over multiple groups.
}
