% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psmodel_specification}
\alias{psmodel_specification}
\title{Specification of propensity score models.}
\format{
A \code{data.frame}. The first two columns are identifiers for an
analysis. The remaining columns are potential
covariates for the propensity score model: a variable is included if it
has a value of 1 and is excluded if it has a value of 0.
}
\usage{
psmodel_specification
}
\description{
A lookup table used to determine the specification of the propensity score
model for each analysis.
}
\examples{
ecmeta.nsclc::psmodel_specification
}
\seealso{
\code{\link[=add_ps_formula]{add_ps_formula()}}, \code{\link[=add_impute_formula]{add_impute_formula()}}
}
\keyword{datasets}
