% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psweight.R
\name{psweight}
\alias{psweight}
\alias{psweight.ps}
\alias{psweight.ps_mi}
\alias{psweight.list}
\title{Propensity score weighting}
\usage{
psweight(object, ...)

\method{psweight}{ps}(
  object,
  methods = c("iptw_att", "iptw_att_trim", "match_nearest", "match_nearest_caliper"),
  ...
)

\method{psweight}{ps_mi}(object, ...)

\method{psweight}{list}(object, progress = FALSE, outfile = "psweight_out", ...)
}
\arguments{
\item{object}{An object of the appropriate class containing the propensity
score to match on and other relevant information.}

\item{...}{Additional arguments to pass to \code{\link[Matching:GenMatch]{Matching::GenMatch()}}.}

\item{methods}{The propensity score methods to use. Can select any combination
of the available methods. The default is to use IPTW-ATT weighting (with
and without trimming) and nearest neighbor matching (with and without a
caliper). Genetic matching can be employed using the options \code{match_genetic}
and \code{match_genetic_caliper}.}

\item{progress}{Logical. If \code{TRUE}, the the iteration number from the list
is printed to the screen and save to the outfile.}

\item{outfile}{File path when \code{progress = TRUE}.}
}
\value{
A \code{dplyr::tibble} is always returned that contains (i) each
variable included in \code{data}, (ii) the columns \code{ps} and \code{logit_ps} and for the
propensity score and logit of the propensity score, respectively,
(iii) the column \code{weight} for the weight generated, (iv) the column \code{method}
for the propensity score method used, (v) the column \code{exclude} denoting
whether patients were excluded using either trimming or a caliper, and (vi)
the column \code{method_type} for the type of propensity score method (weighting,
matching, or none).

The returned object always inherits from \code{\link[dplyr:reexports]{dplyr::tibble}} but the class
depends on \code{object}. If \code{object} is of class \code{\link{ps_fit}}, then there is one row
for each patient and the returned object is of class \code{\link{psweight}}.  If
\code{object} is of class \code{\link{ps_fit_mi}}, then an additional column \code{imp} is added
for the imputation number and an object of class \code{psweight_mi} is returned. Finally,
if \code{object} is of class \code{\link{list}}, then an additional column \code{grp} is added and
an object of class \code{grouped_psweight} or \code{grouped_psweight_mi} is returned.
}
\description{
A generic function used to weight data using estimated propensity scores. Six
separate propensity score methods can be employed: (i) inverse probability weighting
for the average treatment effect on the treated (IPTW-ATT weighting), (ii)
IPTW-ATT weighting where controls with predicted propensity scores less than
the 1st percentile or greater than the 99th percentile are trimmed (i.e., excluded),
(iii) 1:1 nearest neighbor matching without a caliper, (iv) 1:1 nearest
neighbor matching with a caliper (of 0.25 standard deviations of the
linear propensity score), (v) 1:1 genetic matching without a caliper,
and (vi) 1:1 genetic matching with a caliper. Matching is always performed
without replacement unless the number of controls is less then the number of
treated (in which case it is performed with replacement). Unadjusted
(i.e., unweighted) results are also always returned.
}
