% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rbind_list}
\alias{rbind_list}
\title{Row bind data frame like objects}
\usage{
rbind_list(l, id, integer_id = FALSE)
}
\arguments{
\item{l}{A list containing the objects of interest, which must all be of the
same class and inherit from \code{dplyr::tibble}.}

\item{id}{Identifier for for the returned object used to group results.
Passed to the \code{.id} argument of \code{\link[dplyr:bind]{dplyr::bind_rows()}}.}

\item{integer_id}{If \code{TRUE}, then an intetger sequence is used for \code{id}; if
\code{FALSE}, then the names of \code{l} are used.}
}
\value{
A single object of class \code{"grouped_{class}"}
where \code{{class}} is the class of element in \code{l}. The returned object also
inherits from \code{dplyr::tibble} and is grouped by \code{id}.
}
\description{
Row bind a list of objects that inherit from \code{\link[dplyr:reexports]{dplyr::tibble}} to create
a single object grouped by id. Objects in the list must currently be of
class \code{\link{psweight}}, \code{\link{psweight_mi}}, \code{\link{surv_ate}}, or \code{\link{marginal_survival}}.
}
