% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd}
\alias{smd}
\alias{smd.grouped_psweight_mi}
\title{Standardized mean differences}
\usage{
smd(object, ...)

\method{smd}{grouped_psweight_mi}(object, x_vars = NULL)
}
\arguments{
\item{object}{An object of the appropriate class.}

\item{x_vars}{A character vector containing the covariates used to check
for balance. These must be columns in \code{object}. If \code{NULL}, then variables
that were used to fit the propensity score model(s) are used.}
}
\value{
A \code{smd} object that inherits from \code{\link[dplyr:reexports]{dplyr::tibble}}. The following
columns are always returned:
\describe{
\item{var}{A variable - either a covariate or the propensity score.}
\item{method}{The propensity score method.}
\item{mean_control}{The (weighted) mean among the controls.}
\item{mean_treat}{The (weighted) mean among the treated.}
\item{sd_treat}{The standard deviation among the treated.}
\item{smd}{The standardized mean difference.}
}
In addition, a column for the group included if \code{object} is grouped and the
column \code{imp} is included if \code{object} was instantiated from a propensity
score fit using multiply imputed datasets.
}
\description{
Compute standardized mean differences (SMDs) between the treated and control
subjects. Computed as \eqn{(\mu_t - \mu_c)/\sigma} where \eqn{\mu_t} is the
weighted mean among the treated subjects, \eqn{\mu_c} is the weighted mean
among the control subjects, and \eqn{\sigma} is the standard deviation in
the unweighted sample (among the treated subjects only since we are interested
in the average treatment effect for the treated).
}
