% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smd.R
\name{smd_list}
\alias{smd_list}
\title{Standardized mean difference list}
\usage{
smd_list(...)
}
\arguments{
\item{...}{Objects, possibly named.}
}
\value{
An object of class \code{smd_list}.
}
\description{
A list of objects of class \code{\link{smd}}.
}
