% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surv_ate.R
\name{surv_ate}
\alias{surv_ate}
\alias{map_surv_ate}
\title{Average treatment effects for survival data}
\usage{
surv_ate(
  object,
  ydata,
  response,
  id,
  double_adjustment = FALSE,
  ps_stratification = FALSE,
  n_strata = 10
)

map_surv_ate(
  object_list,
  ydata_list,
  response,
  id,
  double_adjustment = FALSE,
  grp_id,
  integer_grp_id = FALSE
)
}
\arguments{
\item{object, object_list}{Either an object of class \code{\link{psweight}} or \code{\link{psweight_mi}} or
a list of such objects.}

\item{ydata, ydata_list}{Either a data frame containing the response variables
for survival modeling or a list of such objects. Both \code{object} and \code{ydata}
must contain the \code{id} variable.}

\item{response}{A string containing the response for a survival model.
This is the let hand side of a \code{~} operator.}

\item{id}{The name of the column in \code{object} and \code{ydata} identifying each
patient.}

\item{double_adjustment}{If \code{TRUE}, then covariates (in addition to treatment
assignment) are used in the Cox models.}

\item{ps_stratification}{If \code{TRUE}, then models are fit that stratify on
values of the propensity score, with the number of strata determined by
\code{n_strata}. Default is \code{FALSE}.}

\item{n_strata}{Number of strata to divide subjects into based on values of
the propensity score.}
}
\value{
A \code{\link[dplyr:reexports]{dplyr::tibble}} with one row for each \code{method} in \code{object}.
Each row contains the columns: \code{method} (the propensity score method),
\code{fit} (\code{tibble} containing the fitted Cox model),
\code{loghr} (\code{tibble} containing estimates of the log hazard ratio), and \code{surv} (A
\code{tibble} containing estimated survival probabilities for both treatment
and control).
}
\description{
Estimate average treatment effects (ATE) from survival data. A Cox proportional
hazards model (\code{\link[survival:coxph]{survival::coxph()}}) is used to estimate hazard ratios and
Kaplan-Meier estimators (\code{survival::survfit.formula()}) are used to estimate
survivor functions. \code{surv_ate()} estimates ATEs for a single group
and \code{map_surv_ate} iterates over groups and estimates ATEs for each group.
Confidence intervals for hazard ratios are estimated by clustering on \code{id} using
the \code{cluster} option in \code{coxph()}. Estimates and confidences intervals are pooled
using Rubin's rule when \code{object} contains weights from multiply imputed datasets.
}
