% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update_ate}
\alias{update_ate}
\alias{update.surv_ate}
\alias{update.grouped_surv_ate}
\alias{update.grouped_ps_surv}
\title{Update treatment effect estimates}
\usage{
\method{update}{surv_ate}(object, ...)

\method{update}{grouped_surv_ate}(object, ...)

\method{update}{grouped_ps_surv}(object, ...)
}
\arguments{
\item{object}{An object of the appropriate class.}
}
\value{
An updated version of \code{object}.
}
\description{
Update models used to estimate average treatment effects (ATEs) with survival data.
This will re-estimate the function used to create \code{object} and is helpful for
performing sensitivity analyses where only a few arguments need to be modified. It
does this by extracting the original function "call".
}
